Preparing structure for composer, minor fixes

+++ core/CardinityUtils.php
@@ -8,6 +8,6 @@ class CardinityUtils
 
     public static function formatAmount($amount)
     {
-        return number_format($amount, 2, '.', '');
+        return (float) number_format($amount, 2, '.', '');
     }
 }
